Add practical example to explainer.
Fix #12

This does not meet @gibson042's requirement:

> but does not demonstrate any use case in which a potential attacker
> has the ability to provide arguments to a sensitiveOperation
> function

but as explained on the issue, I think that's the wrong standard.

If we can assume some mechanism to solve *provisioning*, getting
a sensitiveOperation to a tag function without providing it to
all the tag function's potential callers, then an unbypassable
isTemplateObject check can provide value.

Trusted Types has provisioning machinery, so the example uses that.